Added user verification and admin verification

+++ user_remove_funcs.php
@@ -1,11 +1,12 @@
 <?
 include_once('connection.php');
+include_once('user_get_funcs.php');
 
 function do_leave_group($email, $group_id){
 	// Removes user from group by deleting membership
 	$creator = do_get_creator($group_id);
-
-	if($email != $creator){
+	$admin = user_is_admin($email);
+	if($email != $creator || $admin){
 		$sql = "
 			DELETE FROM `membership`
 			WHERE '$email' = email AND '$group_id' = group_id
@@ -15,7 +16,7 @@ function do_leave_group($email, $group_id){
 			mysql_query('ROLLBACK');
 			return FALSE;
 		}
-		else{
+		else {
 			return TRUE;
 		}
 	}
@@ -31,19 +32,22 @@ function do_remove_message($email, $message_id)
 	//function do_remove_message($email, $message_id, $creator, $admin)
 	// Deletes messages from the database
 	//if email matches the creator OR if admin
-	//if($email == $creator || $admin = true)
-	$sql = "
-		DELETE FROM `messages`
-		WHERE '$email' = creator AND'$message_id' = message_id
-	";
-	$result = mysql_query($sql);
-	if (!$result) {
-		mysql_query('ROLLBACK');
-		return FALSE;
-	}
-	else
-	{
-		return TRUE;
+	$creator = do_get_message_creator($message_id);
+	$admin = user_is_admin($email);
+	if($email == $creator || $admin) {
+		$sql = "
+			DELETE FROM `messages`
+			WHERE '$email' = creator AND'$message_id' = message_id
+		";
+		$result = mysql_query($sql);
+		if (!$result) {
+			mysql_query('ROLLBACK');
+			return FALSE;
+		}
+		else
+		{
+			return TRUE;
+		}
 	}
 }
 
@@ -52,19 +56,24 @@ function do_remove_file($email, $file_name_path)
 {//function do_remove_file($email, $file_name_path, $creator, $admin)
 	// Deletes files from the database
 	//if email matches the creator OR if admin
-	//if($email == $creator || $admin = true)
-	$sql = "
-		DELETE FROM `file`
-		WHERE '$email' = creator AND '$file_name_path' = file_name_path
-	";
-	$result = mysql_query($sql);
-	if (!$result) {
-		mysql_query('ROLLBACK');
-		return FALSE;
-	}
-	else
+	$creator = do_get_file_creator($file_name_path);
+	$admin = user_is_admin($email);
+	if($email == $creator || $admin)
 	{
-		return TRUE;
+		$sql = "
+			DELETE FROM `file`
+			WHERE '$email' = creator AND '$file_name_path' = file_name_path
+		";
+		$result = mysql_query($sql);
+		if (!$result) 
+		{
+			mysql_query('ROLLBACK');
+			return FALSE;
+		}
+		else 
+		{
+			return TRUE;
+		}
 	}
 }
